requirements.txt
discord.py
python-dotenv
requests
beautifulsoup4
groq
flask